requirements.txt
requests==2.25.1
pip-check-reqs==2.2.0
django==2.0
GitPython==3.1.14